+++ src/DevIO.Business/Services/ProdutoService.cs
@@ -1,5 +1,6 @@
 ﻿using DevIO.Business.Interfaces;
 using DevIO.Business.Models;
+using DevIO.Business.Models.Validations;
 
 namespace DevIO.Business.Services
 {
@@ -14,16 +15,32 @@ namespace DevIO.Business.Services
 
         public async Task Adicionar(Produto produto)
         {
+            if (!ExecutarValidacao(new ProdutoValidation(), produto)) return;
+
             await _produtoRepository.Adicionar(produto);
         }
 
         public async Task Atualizar(Produto produto)
         {
+            if (!ExecutarValidacao(new ProdutoValidation(), produto)) return;
+
+            if (!(await _produtoRepository.Buscar(f => f.Id == produto.Id)).Any())
+            {
+                Notificar("Produto não encontrado");
+                return;
+            }
+
             await _produtoRepository.Atualizar(produto);
         }
 
         public async Task Remover(Guid id)
         {
+            if (!(await _produtoRepository.Buscar(f => f.Id == id)).Any())
+            {
+                Notificar("Produto não encontrado");
+                return;
+            }
+
             await _produtoRepository.Remover(id);
         }
